refactor: better file managing
