+++ src/enums/pages.ts
@@ -1,0 +1,5 @@
+export enum Pages {
+  PROJECTS,
+  ABOUT_ME,
+  CONTACT,
+}
